[+] DisplayCopiableText component
